requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
flask-talisman==1.1.0
Flask-WTF==1.1.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.3
mailersend==0.5.6
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.2.1
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
